fixing role in member list
