+++ src/components/CurrentColony/settings/informations/informations.css
