+++ src/components/CurrentColony/settings/informations/informations.tsx
